rm special component for extrusion layer
